requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
dashscope
chromadb
tiktoken
wikipedia-api
python-dotenv
pypdf
PyMuPDF
pillow
IPython
matplotlib
beautifulsoup4
unstructured
langchain-unstructured
faiss-cpu
pinecone
langchain_pinecone
pinecone-text
pinecone-notebooks
rank_bm25
nltk
langchain-chroma
langchainhub
langgraph
graph-rag-example-helpers
langchain-graph-retriever
dingtalk-stream
serpapi
langchain-qdrant
qdrant-client
google-search-results
redis